split the app.ts into different files
