+++ src/colortype.ts
@@ -1,0 +1,9 @@
+export type ColorType = {
+  name: string;
+  color: string;
+};
+
+export const initialColor: ColorType = {
+  name: "beige",
+  color: "rgb(245,245,220)",
+};
